+++ Sources/VCore/Views/SwiftUI/Touch Sensitive Container/TouchSensitiveContainerState.swift
@@ -1,0 +1,12 @@
+//
+//  TouchSensitiveContainerState.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 17.09.23.
+//
+
+import Foundation
+
+// MARK: - Touch Sensitive Container Internal State
+/// Enumeration that represents state, such as `enabled`, `pressed`, or `disabled`.
+public typealias TouchSensitiveContainerInternalState = GenericState_EnabledPressedDisabled